refactor(utils): changed int type to long

+++ plugin/src/main/java/team/aquatic/betterjoin/utils/Utils.java
@@ -53,7 +53,7 @@ public class Utils {
 	public static void showTitle(
 		 @NotNull Player player,
 		 @NotNull String title, @NotNull String subtitle,
-		 int fadeIn, int stay, int fadeOut
+		 long fadeIn, long stay, long fadeOut
 	) {
 		Objects.requireNonNull(player, "Player is null.");
 		Validate.notEmpty(title, "The title is empty.");
